deleted src/render.js, refact
