--- src/render.js
@@ -1,22 +1,0 @@
-import App from "./App";
-import { addPost } from "./Redux/state";
-import { updateNewPostText } from "./Redux/state";
-import { BrowserRouter } from "react-router-dom";
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-let rerenderEntireThree = (state) => {
-  root.render(
-    <BrowserRouter>
-      <App
-        state={state}
-        addPost={addPost}
-        updateNewPostText={updateNewPostText}
-      />
-    </BrowserRouter>
-  );
-};
-
-export default rerenderEntireThree;
